Electronic Accessories: An e-girls best friend
Thanks to Undin all the electronic accessories have been added.

Cleaned up the code for generating RFID tags a bit.

Tried to find all the variables that are meant to only be in function
and add “var” to the front of them, so they don’t effect the global
variables. I still seem to have a global variable for device somewhere
that I can’t seem to find.

Made it so that communication devices without ratings have their rating
up and down buttons disabled.